+++ WebContent/css/professoradd.css
@@ -1,4 +1,3 @@
-/* style.css */
 
 body {
     font-family: Arial, sans-serif;
@@ -86,4 +85,3 @@ input[type="button"]:hover {
     display: none;
     margin-top: 5px;
 }
-
