+++ src/Dao/MemberDAO.java
@@ -102,6 +102,7 @@ public class MemberDAO {
 
 	}
 	
+	//===============================================================================================
 	//로그인한 회원 아이디를 이용해 회원정보 조회
 		//이유 : 글작성 화면에 조회한 회원정보를 보여주기 위해 
 		public MemberVo memberOne(String reply_id_) {
@@ -129,193 +130,4 @@ public class MemberDAO {
 			return vo;//BoardService로 리턴(보고)
 		}
 
-	// =========================================================================================================
-	// 학생 등록
-	public int insertStudent(MemberVo studentUser) {
-		int result = 0;
-		try {
-			con = getConnection();
-			query = "INSERT INTO user (user_id, user_pw, user_name, birthDate, gender, address, phone, email, role) "
-					+ "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)";
-			pstmt = con.prepareStatement(query);
-			pstmt.setString(1, studentUser.getUser_id());
-			pstmt.setString(2, studentUser.getUser_pw());
-			pstmt.setString(3, studentUser.getUser_name());
-			pstmt.setDate(4, studentUser.getBirthDate());
-			pstmt.setString(5, studentUser.getGender());
-			pstmt.setString(6, studentUser.getAddress());
-			pstmt.setString(7, studentUser.getPhone());
-			pstmt.setString(8, studentUser.getEmail());
-			pstmt.setString(9, studentUser.getRole());
-			result = pstmt.executeUpdate();
-
-			query = "INSERT INTO student_info (student_id, user_id, majorcode, grade, admission_date, status) "
-					+ "VALUES (?, ?, ?, ?, ?, ?)";
-			pstmt = con.prepareStatement(query);
-			pstmt.setString(1, studentUser.getStudent_id());
-			pstmt.setString(2, studentUser.getUser_id());
-			pstmt.setString(3, studentUser.getMajorcode());
-			pstmt.setInt(4, studentUser.getGrade());
-			pstmt.setDate(5, studentUser.getAdmission_date());
-			pstmt.setString(6, studentUser.getStatus());
-			result = pstmt.executeUpdate();
-		} catch (SQLException e) {
-			System.err.println("학생등록오류: " + e.getMessage());
-			e.printStackTrace();
-		} finally {
-			closeResource();
-		}
-		return result;
-	}
-
-	// =========================================================================================================
-	// 전체 학생 조회
-	public List<MemberVo> getAllStudents() {
-		List<MemberVo> students = new ArrayList<>();
-		String query = "SELECT s.student_id, u.user_id, u.user_name, s.majorcode, s.grade, s.admission_date, s.status "
-				+ "FROM student_info s JOIN user u ON s.user_id = u.user_id";
-		try {
-			con = getConnection();
-			pstmt = con.prepareStatement(query);
-			rs = pstmt.executeQuery();
-
-			while (rs.next()) {
-				MemberVo student = new MemberVo();
-				student.setStudent_id(rs.getString("student_id"));
-				student.setUser_id(rs.getString("user_id"));
-				student.setUser_name(rs.getString("user_name"));
-				student.setMajorcode(rs.getString("majorcode"));
-				student.setGrade(rs.getInt("grade"));
-				student.setAdmission_date(rs.getDate("admission_date"));
-				student.setStatus(rs.getString("status"));
-				students.add(student);
-			}
-		} catch (SQLException e) {
-			System.err.println("전체 학생 목록 조회 중 예외 발생: " + e.getMessage());
-			e.printStackTrace();
-		} finally {
-			closeResource();
-		}
-		return students;
-	}
-
-	// =========================================================================================================
-	// 특정 학생 검색해서 조회이긴한데 구현은 안됨. 하지만 있어도 상관없기에 일단 넣어둠 나중에 추가할 수 있으면 추가하면 될 듯
-	public MemberVo getStudentById(String userId) {
-		MemberVo student = null;
-		String query = "SELECT u.user_id, u.user_pw, u.user_name, u.birthDate, u.gender, u.address, u.phone, "
-				+ "u.email, u.role, s.student_id, s.majorcode, s.grade, s.admission_date, s.status "
-				+ "FROM user u JOIN student_info s ON u.user_id = s.user_id WHERE u.user_id = ?";
-		try {
-			con = getConnection();
-			pstmt = con.prepareStatement(query);
-			pstmt.setString(1, userId);
-			rs = pstmt.executeQuery();
-
-			if (rs.next()) {
-				student = new MemberVo();
-				student.setUser_id(rs.getString("user_id"));
-				student.setUser_pw(rs.getString("user_pw"));
-				student.setUser_name(rs.getString("user_name"));
-				student.setBirthDate(rs.getDate("birthDate"));
-				student.setGender(rs.getString("gender"));
-				student.setAddress(rs.getString("address"));
-				student.setPhone(rs.getString("phone"));
-				student.setEmail(rs.getString("email"));
-				student.setRole(rs.getString("role"));
-				student.setStudent_id(rs.getString("student_id"));
-				student.setMajorcode(rs.getString("majorcode"));
-				student.setGrade(rs.getInt("grade"));
-				student.setAdmission_date(rs.getDate("admission_date"));
-				student.setStatus(rs.getString("status"));
-			}
-		} catch (SQLException e) {
-			System.err.println("특정 학생 조회 중 예외 발생: " + e.getMessage());
-			e.printStackTrace();
-		} finally {
-			closeResource();
-		}
-		return student;
-	}
-
-	// =========================================================================================================
-	// 학생 정보 수정
-	public boolean updateStudent(MemberVo student) {
-		boolean isUpdated = false;
-		String query = "UPDATE user u JOIN student_info s ON u.user_id = s.user_id SET "
-				+ "u.user_pw = ?, u.user_name = ?, u.birthDate = ?, u.gender = ?, "
-				+ "u.address = ?, u.phone = ?, u.email = ?, u.role = ?, "
-				+ "s.grade = ?, s.admission_date = ?, s.status = ? " + "WHERE s.student_id = ?";
-		try {
-			con = getConnection();
-			pstmt = con.prepareStatement(query);
-			pstmt.setString(1, student.getUser_pw());
-			pstmt.setString(2, student.getUser_name());
-			pstmt.setDate(3, student.getBirthDate());
-			pstmt.setString(4, student.getGender());
-			pstmt.setString(5, student.getAddress());
-			pstmt.setString(6, student.getPhone());
-			pstmt.setString(7, student.getEmail());
-			pstmt.setString(8, student.getRole());
-			pstmt.setInt(9, student.getGrade());
-			pstmt.setDate(10, student.getAdmission_date());
-			pstmt.setString(11, student.getStatus());
-			pstmt.setString(12, student.getStudent_id());
-
-			int rowsUpdated = pstmt.executeUpdate();
-			isUpdated = rowsUpdated > 0;
-		} catch (SQLException e) {
-			e.printStackTrace();
-		} finally {
-			closeResource();
-		}
-		return isUpdated;
-	}
-
-	// =========================================================================================================
-	// 학생 정보 삭제
-	public boolean deleteStudent(String studentId) {
-		boolean isDeleted = false;
-		String selectQuery = "SELECT user_id FROM student_info WHERE student_id = ?";
-		String deleteStudentInfoQuery = "DELETE FROM student_info WHERE student_id = ?";
-		String deleteUserQuery = "DELETE FROM user WHERE user_id = ?";
-
-		try {
-			con = getConnection();
-			con.setAutoCommit(false);
-			String userId = null;
-
-			try (PreparedStatement pstmtSelect = con.prepareStatement(selectQuery)) {
-				pstmtSelect.setString(1, studentId);
-				ResultSet rs = pstmtSelect.executeQuery();
-				if (rs.next()) {
-					userId = rs.getString("user_id");
-				}
-			}
-
-			if (userId != null) {
-				try (PreparedStatement pstmtDeleteStudentInfo = con.prepareStatement(deleteStudentInfoQuery)) {
-					pstmtDeleteStudentInfo.setString(1, studentId);
-					pstmtDeleteStudentInfo.executeUpdate();
-				}
-				try (PreparedStatement pstmtDeleteUser = con.prepareStatement(deleteUserQuery)) {
-					pstmtDeleteUser.setString(1, userId);
-					pstmtDeleteUser.executeUpdate();
-				}
-				con.commit();
-				isDeleted = true;
-			}
-		} catch (SQLException e) {
-			System.err.println("학생 삭제 중 예외 발생: " + e.getMessage());
-			e.printStackTrace();
-			try {
-				con.rollback();
-			} catch (SQLException rollbackEx) {
-				rollbackEx.printStackTrace();
-			}
-		} finally {
-			closeResource();
-		}
-		return isDeleted;
-	}
 }// MemberDAO 클래스
